+++ pydantic/__init__.py
@@ -3,5 +3,6 @@ from .env_settings import BaseSettings
 from .exceptions import *
 from .fields import Required
 from .main import BaseModel
+from .parse import Protocol
 from .types import *
 from .version import VERSION